Move `Hero` to `Home` module
